# Allow rewriting in parallel