--- a/wea/src/lib.rs
+++ b/wea/src/lib.rs
@@ -6,7 +6,6 @@
 //! To import some `X` from inside this library, it may sometimes be necessary
 //! to add a `pub use::X` to `main.rs`.
 #![allow(clippy::new_without_default)]
-#![allow(clippy::arc_with_non_send_sync)]
 
 mod op;
 mod wea_to_mlir;