--- a/xrcf/Cargo.toml
+++ b/xrcf/Cargo.toml
@@ -18,9 +18,10 @@ harness = false
 [dependencies]
 anyhow = "1.0.86"
 clap = { version = "4.5", features = ["derive"] }
+parking_lot = { version = "0.12", features = ["deadlock_detection"] }
+rayon = "1.10"
 tracing = "0.1"
 tracing-subscriber = "0.3"
-parking_lot = { version = "0.12", features = ["deadlock_detection"] }
 wasmtime = { version = "28.0.0", features = ["cranelift"] }
 
 [dev-dependencies]