## Benchmark

The benchmark can be executed via `cargo bench`. During the benchmark, xrcf rewrites the following function that is repeated 100 times in a file
```mlir
func.func @add_one(%arg0 : i32) -> i32 {
    %0 = arith.constant 1 : i32
    %1 = arith.addi %0, %arg0 : i32
    return %1 : i32
}
```
to
```mlir
llvm.func @add_one(%arg0 : i32) -> i32 {
    %0 = llvm.mlir.constant(1 : i32) : i32
    %1 = llvm.add %0, %arg0 : i32
    llvm.return %1 : i32
}
```

## Result

Before:
```raw
time:   [47.268 ms 47.330 ms 47.377 ms]
```

After:
```raw
time:   [24.160 ms 24.355 ms 24.558 ms]
```

The overhead of `DefaultTester::verify` is negligible here because the rewrite with one repeat takes 0.4 ms.

## Discussion

Closes https://github.com/rikhuijzer/xrcf/issues/37. 

This now closes the discussion in https://github.com/rikhuijzer/xrcf/pull/41. In summary, query-based compilers are nice, but not the best fit for this project. Many compilers such as C, Zig ([source](https://youtu.be/IroPQ150F6c?t=1936)), Mojo ([source](https://youtu.be/yuSBEXkjfEA)), and Cranelift are not query-based but instead compile files in parallel and then link the files sequentially. (Note that linking means matching up, for example, function definitions with function calls.)